backend/requirements.txt
# ---- FastAPI
fastapi==0.111.0
uvicorn==0.30.1
httptools==0.6.4

# ---- DB & Vector
psycopg2-binary==2.9.9
pgvector==0.2.5

# ---- LangChain stack
langchain==0.3.28
langchain-community==0.3.31
langchain-core==0.3.83
langchain-google-genai==2.0.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.11
flashrank
sentence-transformers
requests

# ---- Loaders & utils
langchain-docling
pypdf
python-dotenv
python-multipart

# ---- Whisper
faster-whisper
soundfile

# ---- OCR
pytesseract
Pillow

# ---- Utilities
tqdm

# ---- PDF
PyMuPDF
cryptography>=3.1

# ---- Word
python-docx

# ---- authentication
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]
email-validator
